Various housing fixes, addition of exitRange handling for housing zones

+++ src/world/Network/Handlers/PacketCommandHandler.cpp
@@ -694,7 +694,7 @@ void Sapphire::Network::GameConnection::commandHandler( const Packets::FFXIVARR_
     {
       auto& housingMgr = Service< HousingMgr >::ref();
 
-      housingMgr.sendWardLandInfo( player, static_cast< uint8_t >( data.Arg1 ), static_cast< uint8_t >( data.Arg0 ) );
+      housingMgr.sendWardLandInfo( player, static_cast< uint8_t >( data.Arg1 ), static_cast< uint16_t >( data.Arg0 ) );
 
       break;
     }
